Adds waning effect to the vaccine
The waning period is chosen from an exponential distribution.
In the SimControl file the variable VaccineWanePeriod specifies the
mean of this effect.

Every day vaccinated individuals are checked to see if the vaccine
wanes that day. If it does, they become unvaccinated.

Risk of infection was included in the infect procedure in Human.cpp

I hope that I fixed some indentation problems.