adds  the reports for discrete ages: 0,1,2,3,etc.
This update adds the reports for discrete ages: 0,1,2,3,etc. with a
couple of additional functions in the Report class. There are still
some optimizations to do, mostly if we want to create reports
asynchronously (Which we don’t right now).

This update also fixes a bug in the updateReports() routine. Everyday,
human variables were being reset to 0.