Fix of vaccine-induced immunity
This should fix the issue with the immunity acquired by vaccination.
The variable vaccine_immunity is used to check if a person has temporal
immunity from vaccination. It differs from immunityTemp or
immunityPerm, which are natural immunity.

To check anything vaccine-immune related used the methods with Vax in
them. like isImmuneVax() or setVaxImmunity().