First submission of the GSK trial branch
The recruitment class randomly selects participants to enroll in the
placebo or vaccine arm of the study.

Vaccine settings are read from an external file.

Trial settings are read from an external file.

Multiple vaccine profiles can be created and the IDs can be specified
in the trial or SimControl file.