Vaccine class
A new vaccine class has been defined. It should be a basic class
containing the parameters of each simulated vaccine.

Right now the recruitment takes a map of vaccines instead of a pointer.
It was causing segmentation faults. I should probably try to fix that
soon. But it works passing the map instead of the pointer.